Refactor Serializers into subpackage
