+++ src/main/java/data_access/serialization/QuestionSerializer.java
@@ -1,4 +1,4 @@
-package data_access;
+package data_access.serialization;
 
 import entity.Question;
 import org.json.JSONObject;
