+++ src/main/java/data_access/serialization/ResponseSerializer.java
@@ -1,4 +1,4 @@
-package data_access;
+package data_access.serialization;
 
 import entity.Answer;
 import entity.AnswerType;
